feat(questions): add styles and refactor imports for GroupButton and GroupSelectionHeader components
