+++ app/(tabs)/questions/GroupSelectionHeader.tsx
@@ -1,6 +1,6 @@
 import { ThemedText } from '@/components/ThemedText'
 import { View } from 'react-native'
-import { styles } from './styles'
+import { styles } from './_styles'
 
 export function GroupSelectionHeader() {
   return (
