+++ app/(tabs)/questions/_styles.ts
