+++ app/features/solve-problems/hooks/useGroupSelection.ts
